+++ demo/app/models/follow.rb
@@ -1,2 +1,3 @@
-class Follow < Socialization::ActiveRecordStores::FollowStore
+class Follow < Socialization::ActiveRecordStores::Follow
+  self.table_name = 'follows'
 end