+++ lib/generators/socialization/templates/active_record/model_follow.rb
@@ -1,2 +1,2 @@
-class Follow < Socialization::ActiveRecordStores::FollowStore
+class Follow < Socialization::ActiveRecordStores::Follow
 end