+++ lib/generators/socialization/templates/redis/model_follow.rb
@@ -1,2 +1,2 @@
-class Follow < Socialization::RedisStores::FollowStore
+class Follow < Socialization::RedisStores::Follow
 end