+++ lib/socialization/stores/active_record/mention.rb
@@ -1,14 +1,29 @@
-require File.expand_path(File.dirname(__FILE__)) + '/base'
-
 module Socialization
-  module RedisStores
-    class MentionStore < Socialization::RedisStores::Base
+  module ActiveRecordStores
+    class Mention < ActiveRecord::Base
+      belongs_to :mentioner,   :polymorphic => true
+      belongs_to :mentionable, :polymorphic => true
+
+      scope :mentioned_by, lambda { |mentioner| where(
+        :mentioner_type   => mentioner.class.table_name.classify,
+        :mentioner_id     => mentioner.id)
+      }
+
+      scope :mentioning,   lambda { |mentionable| where(
+        :mentionable_type => mentionable.class.table_name.classify,
+        :mentionable_id   => mentionable.id)
+      }
+
+      @@after_create_hook = nil
+      @@after_destroy_hook = nil
+
       class << self
         def mention!(mentioner, mentionable)
           unless mentions?(mentioner, mentionable)
-            Socialization.redis.sadd generate_mentioners_key(mentioner, mentionable), mentioner.id
-            Socialization.redis.sadd generate_mentionables_key(mentioner, mentionable), mentionable.id
-
+            self.create! do |mention|
+              mention.mentioner = mentioner
+              mention.mentionable = mentionable
+            end
             call_after_create_hook(mentioner, mentionable)
             mentioner.touch if [:all, :mentioner].include?(touch) && mentioner.respond_to?(:touch)
             mentionable.touch if [:all, :mentionable].include?(touch) && mentionable.respond_to?(:touch)
@@ -20,9 +35,7 @@ module Socialization
 
         def unmention!(mentioner, mentionable)
           if mentions?(mentioner, mentionable)
-            Socialization.redis.srem generate_mentioners_key(mentioner, mentionable), mentioner.id
-            Socialization.redis.srem generate_mentionables_key(mentioner, mentionable), mentionable.id
-
+            mention_for(mentioner, mentionable).destroy_all
             call_after_destroy_hook(mentioner, mentionable)
             mentioner.touch if [:all, :mentioner].include?(touch) && mentioner.respond_to?(:touch)
             mentionable.touch if [:all, :mentionable].include?(touch) && mentionable.respond_to?(:touch)
@@ -33,40 +46,58 @@ module Socialization
         end
 
         def mentions?(mentioner, mentionable)
-          Socialization.redis.sismember generate_mentioners_key(mentioner, mentionable), mentioner.id
+          !mention_for(mentioner, mentionable).empty?
         end
 
         # Returns an ActiveRecord::Relation of all the mentioners of a certain type that are mentioning mentionable
         def mentioners_relation(mentionable, klass, opts = {})
-          ids = mentioners(mentionable, klass, :pluck => :id)
-          klass.where('id IN (?)', ids)
+          rel = klass.where(:id =>
+            self.select(:mentioner_id).
+              where(:mentioner_type => klass.table_name.classify).
+              where(:mentionable_type => mentionable.class.to_s).
+              where(:mentionable_id => mentionable.id)
+          )
+
+          if opts[:pluck]
+            rel.pluck(opts[:pluck])
+          else
+            rel
+          end
         end
 
         # Returns all the mentioners of a certain type that are mentioning mentionable
         def mentioners(mentionable, klass, opts = {})
-          if opts[:pluck]
-            Socialization.redis.smembers(generate_mentioners_key(klass, mentionable)).map { |id|
-              id.to_i if id.is_integer?
-            }
+          rel = mentioners_relation(mentionable, klass, opts)
+          if rel.is_a?(ActiveRecord::Relation)
+            rel.all
           else
-            mentioners_relation(mentionable, klass, opts).all
+            rel
           end
         end
 
         # Returns an ActiveRecord::Relation of all the mentionables of a certain type that are mentioned by mentioner
         def mentionables_relation(mentioner, klass, opts = {})
-          ids = mentionables(mentioner, klass, :pluck => :id)
-          klass.where('id IN (?)', ids)
+          rel = klass.where(:id =>
+            self.select(:mentionable_id).
+              where(:mentionable_type => klass.table_name.classify).
+              where(:mentioner_type => mentioner.class.to_s).
+              where(:mentioner_id => mentioner.id)
+          )
+
+          if opts[:pluck]
+            rel.pluck(opts[:pluck])
+          else
+            rel
+          end
         end
 
         # Returns all the mentionables of a certain type that are mentioned by mentioner
         def mentionables(mentioner, klass, opts = {})
-          if opts[:pluck]
-            Socialization.redis.smembers(generate_mentionables_key(mentioner, klass)).map { |id|
-              id.to_i if id.is_integer?
-            }
+          rel = mentionables_relation(mentioner, klass, opts)
+          if rel.is_a?(ActiveRecord::Relation)
+            rel.all
           else
-            mentionables_relation(mentioner, klass, opts).all
+            rel
           end
         end
 
@@ -98,26 +129,8 @@ module Socialization
           self.send(@after_destroy_hook, mentioner, mentionable) if @after_destroy_hook
         end
 
-        def generate_mentioners_key(mentioner, mentionable)
-          raise ArgumentError.new("`mentionable` needs to be an acts_as_mentionable objecs, not a class.") if mentionable.class == Class
-          mentioner_class = if mentioner.class == Class
-            mentioner
-          else
-            mentioner.class
-          end
-
-          "Mentioners:#{mentionable.class}:#{mentionable.id}:#{mentioner_class}"
-        end
-
-        def generate_mentionables_key(mentioner, mentionable)
-          raise ArgumentError.new("`mentioner` needs to be an acts_as_mentioner object, not a class.") if mentioner.class == Class
-          mentionable_class = if mentionable.class == Class
-            mentionable
-          else
-            mentionable.class
-          end
-
-          "Mentionables:#{mentioner.class}:#{mentioner.id}:#{mentionable_class}"
+        def mention_for(mentioner, mentionable)
+          mentioned_by(mentioner).mentioning(mentionable)
         end
       end # class << self